add nationwide-specific export command; don't build CHs for nationwide build
