+++ nationwide_gh_config.yaml
@@ -7,22 +7,24 @@ graphhopper:
   graph.location: transit_data/graphhopper
   routing.ch.disabling_allowed: true
   routing.max_visited_nodes: 1500000
-  graph.flag_encoders: car
+  graph.flag_encoders: car,bike,foot,truck
   prepare.ch.threads: 16
 
-  # Uncomment this if the export was done with turn-restriction-aware contraction hierarchies
-  # prepare.ch.edge_based: edge_and_node
-  # graph.flag_encoders: car|turn_costs=true
-  # graph.bytes_for_flags: 64
-
   # Profiles specifying vehicle and weightings for each mode type.
+  # For nationwide export, save time by not building contraction hierarchies
   profiles:
     - name: car
       vehicle: car
       weighting: fastest
-
-  profiles_ch:
-    - profile: car
+    - name: bike
+      vehicle: bike
+      weighting: fastest
+    - name: foot
+      vehicle: foot
+      weighting: fastest
+    - name: truck
+      vehicle: truck
+      weighting: fastest
 
 server:
   min_threads: 4
