+++ web-bundle/pom.xml
@@ -29,6 +29,12 @@
             <version>2.3.1</version>
         </dependency>
 
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-csv</artifactId>
+            <version>1.4</version>
+        </dependency>
+
         <!--
         <dependency>
             <groupId>com.datadoghq</groupId>
