+++ web/src/main/java/com/graphhopper/http/GraphHopperApplication.java
@@ -18,6 +18,7 @@
 package com.graphhopper.http;
 
 import com.graphhopper.http.cli.ExportCommand;
+import com.graphhopper.http.cli.ExportNationwideCommand;
 import com.graphhopper.http.cli.GtfsLinkMapperCommand;
 import com.graphhopper.http.cli.ImportCommand;
 import com.graphhopper.http.resources.RootResource;
@@ -44,6 +45,7 @@ public final class GraphHopperApplication extends Application<GraphHopperServerC
         bootstrap.addCommand(new ImportCommand());
         bootstrap.addCommand(new GtfsLinkMapperCommand());
         bootstrap.addCommand(new ExportCommand());
+        bootstrap.addCommand(new ExportNationwideCommand());
 
         Map<String, String> resourceToURIMappings = new HashMap<>();
         resourceToURIMappings.put("/src/main/resources/assets/", "/maps/");
